requirements.txt
# Core
numpy
pandas
pydantic

# API + DB
fastapi
uvicorn[standard]
sqlmodel
SQLAlchemy
python-dotenv

# (Optional) LLMs later
transformers

# Visualization
matplotlib
seaborn
